chore: revert "feat: split Geocoder widget in two: Widget & WidgetUI"
This reverts commit 6475f044a29dc8b7d809a42f6c354603b6992225.
